+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: melperri <melperri@student42.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/08 17:28:17 by melperri          #+#    #+#             */
-/*   Updated: 2020/10/08 17:28:35 by melperri         ###   ########.fr       */
+/*   Updated: 2020/10/12 12:47:56 by melperri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
